Remove the optional 'optional' parameter from AddDockerSecrets
To be more inline with Arcus fylosophy of failing fast, it has been decided to remove the parameter that declares whether the presence of the configured directory that contains the docker secrets is optional or not.
When the mounted directory is not present, we will fail when setting up the secretprovider.