fix(az.kv): invalidate secret after az kv secret provider sets new secret (version) (#487)
When a new secret (version) is set via the `KeyVaultSecretProvider`,
then the secret provider implementation should invalidate the secret
back to the secret store cache. This wasn't done, which resulted in
still cached secrets when the secret was retrieved.
This PR fixes this by injecting the secret store context into the
`KeyVaultSecretProvider` and calling the 'invalidate secret'
functionality on it when a new secret (version) is set.

(Functionality not yet released, so not really a public bug.)

Closes #486