docs(v3): rework feature docs for v3 (#485)
Restructure, rewrite and simplify (preview) feature documentation in
preparation for the v3 release. It also includes the local search
(instead of the unstable Algoria) and uses PowerShell as code language
to show installation commands.

Furthermore, the 'common' functionality is located at the secret store
index page and an introduction/getting-started page is added to guide
consumers more into the Arcus Security space.

(Later, a dropdown menu with the available Arcus spaces might be
beneficial as well.)

---------

Co-authored-by: Frederik Gheysels <frederik.gheysels@telenet.be>