fix(hashicorp): use secret mapping in hashicorp vault tests (#500)
Since the secret mapping function was not part of storing the HashiCorp
Vault secret in the tests, it sometimes occurred that the test failed
(randomization of adding/skipping options does that).

This PR adds the optional secret name mapping so it always takes
optional mapping into account when storing HashiCorp Vault secrets.

Closes #499